Implement extract from directory
The Safari browser automatically unzips downloaded zipfiles by default,
so it's convenient to be able to operate on directories too.

Perhaps this functionality should really be in a seperate command, but
this is fairly simple on the implementation side. We'll get features in
now and worry about the interface later. Worked for git, right?

Fixes #4.