Set patool as noninteractive to prevent hanging
This will dismiss all overwrite file prompts, thereby resolving hangs
and conflicts. It's not a terribly pretty solution, but there's no
obvious answer as to which files to keep when they conflict within the
same submission.

If we find some time to put into this in the future, we could instead
check if the files are identical or not, and in the case that they're
different we could just rename one file.

Fixes #7.